feat: add OpenGraph and Twitter card meta tags to results page
